# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LeadEngine by Novenworks - Multi-client lead capture CRM API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: clients
    description: Client/business management
  - name: leads
    description: Lead management
  - name: dashboard
    description: Dashboard statistics
  - name: public
    description: Public lead capture endpoints
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current session
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients (admin sees all, owner sees their own)
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Client deleted
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /clients/{id}/embed-code:
    get:
      operationId: getEmbedCode
      tags: [clients]
      summary: Get HTML embed code snippet for a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Embed code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbedCode"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads with optional filters
      parameters:
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"

  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead status, notes, or last contacted date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leads/capture:
    post:
      operationId: captureLead
      tags: [public]
      summary: Public endpoint to capture a lead from a client website
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadCapture"
      responses:
        "201":
          description: Lead captured successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid API key or slug
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get lead stats for the dashboard
      parameters:
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/recent:
    get:
      operationId: getRecentLeads
      tags: [dashboard]
      summary: Get recent leads for dashboard activity feed
      parameters:
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Recent leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AuthUser:
      type: object
      required: [id, email, role, name]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, owner]
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]

    Client:
      type: object
      required: [id, businessName, slug, ownerEmail, notificationEmail, apiKey, createdAt]
      properties:
        id:
          type: integer
        businessName:
          type: string
        slug:
          type: string
        ownerEmail:
          type: string
        notificationEmail:
          type: string
        apiKey:
          type: string
        websiteUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    ClientInput:
      type: object
      required: [businessName, slug, ownerEmail, notificationEmail]
      properties:
        businessName:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        ownerEmail:
          type: string
        notificationEmail:
          type: string
        websiteUrl:
          type: string

    ClientUpdate:
      type: object
      properties:
        businessName:
          type: string
          minLength: 1
        ownerEmail:
          type: string
        notificationEmail:
          type: string
        websiteUrl:
          type: string

    EmbedCode:
      type: object
      required: [clientId, businessName, slug, apiKey, htmlSnippet]
      properties:
        clientId:
          type: integer
        businessName:
          type: string
        slug:
          type: string
        apiKey:
          type: string
        htmlSnippet:
          type: string

    Lead:
      type: object
      required: [id, clientId, name, email, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        serviceInterest:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        status:
          type: string
          enum: [New, Contacted, Booked, Won, Lost]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastContactedAt:
          type: ["string", "null"]
          format: date-time

    LeadUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [New, Contacted, Booked, Won, Lost]
        notes:
          type: string
        lastContactedAt:
          type: string
          format: date-time

    LeadCapture:
      type: object
      required: [clientSlug, apiKey, name, email]
      properties:
        clientSlug:
          type: string
        apiKey:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        serviceInterest:
          type: string
        message:
          type: string
        source:
          type: string

    DashboardStats:
      type: object
      required: [total, new, contacted, booked, won, lost]
      properties:
        total:
          type: integer
        new:
          type: integer
        contacted:
          type: integer
        booked:
          type: integer
        won:
          type: integer
        lost:
          type: integer