# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LeadEngine by Novenworks - Multi-client lead capture CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from './authUserRole';

export interface AuthUser {
  id: number;
  email: string;
  name: string;
  role: AuthUserRole;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
}